requirements.txt
numpy==1.26.4
Pillow==10.0.1
ping3==4.0.8
PyAutoGUI==0.9.54